+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Unreleased
 ### Added
 - remote grainy permissions
+- field level permission checks when POST/PUT/PATCHing to REST api (#14)
 ### Changed
 - fix #36: Grainy decorators should allow setting of Permissions class to use for permission checks
 ### Removed